ch(refactor): update folder structure
