+++ src/graphql/index.ts
@@ -1,1 +1,1 @@
-export * from './User';
+export * from './user/index';
