+++ src/graphql/user/index.ts
@@ -1,0 +1,3 @@
+export * from './type';
+export * from './query';
+export * from './mutation';
