+++ src/schema.ts
@@ -1,6 +1,6 @@
 import { makeSchema } from 'nexus';
 import { join } from 'path';
-import * as types from './graphql';
+import * as types from './graphql/index';
 
 const schema = makeSchema({
 	types,
